refactor: split game screen overlays, production panel, trade host (#3878)
Extract GameScreen overlay composition, ProductionPanel constants/layouts,
and TradeScreen widget host into library parts to keep screen hosts thin
without changing public APIs or player-facing behavior.

Refs #3878
