+++ app/lib/features/game/screens/game/game_screen.dart
@@ -3,6 +3,7 @@ export 'game_screen_shared.dart';
 import 'dart:async';
 
 import 'package:colonizethis_logic/colonizethis_logic.dart';
+import 'package:colonizethis_map/colonizethis_map.dart' show InitGameMapViewData;
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_app/package_logger.dart';
 import 'package:flame/game.dart' hide Game;
@@ -44,6 +45,7 @@ import '../../../../core/services/turn_resolution/turn_resolution_result_applier
 import '../../flame/overlays/victory_overlay.dart';
 
 part 'game_screen_fallback_next_turn.dart';
+part 'game_screen_overlay_stack.dart';
 
 final _gameScreenLog = packageLogger('logic');
 
@@ -79,126 +81,16 @@ class GameScreen extends ConsumerWidget {
         heraldQueue.isNotEmpty ? heraldQueue.first : null;
     final pendingDiplomacy = ref.watch(pendingDiplomacyProvider);
     final turnResolutionBlocking = ref.watch(turnResolutionBlockingProvider);
-    Widget content = Stack(
-      children: [
-        if (mapViewData != null && game != null)
-          GameMapArea(game: game, mapViewData: mapViewData)
-        else
-          GameWidget(game: ColonizeThisGame()),
-        if (showOverlayButtons) ...[
-          Positioned(
-            left: 16,
-            top: 16,
-            child: CtIconAction(
-              icon: Icons.menu,
-              iconSize: 24,
-              onPressed: () => _showPauseMenu(ref.read(appEventBusProvider)),
-              tooltip: appL10n(context).game_pauseMenu_tooltip,
-            ),
-          ),
-          Positioned(
-            right: 16,
-            top: 16,
-            child: _FallbackNextTurnButton(
-              game: game,
-              turnResolutionBlocking: turnResolutionBlocking,
-            ),
-          ),
-        ],
-        if (game != null && victory != null)
-          VictoryOverlay(
-            game: game,
-            victory: victory,
-            bus: ref.read(appEventBusProvider),
-          ),
-      ],
+    final content = _GameScreenOverlayStack(
+      game: game,
+      mapViewData: mapViewData,
+      victory: victory,
+      showOverlayButtons: showOverlayButtons,
+      showIntro: showIntro,
+      pendingHerald: pendingHerald,
+      pendingDiplomacy: pendingDiplomacy,
+      turnResolutionBlocking: turnResolutionBlocking,
     );
-
-    if (game != null) {
-      content = TribeFirstContactSyncListener(
-        child: GameToUIBusListener(gameId: game.id, child: content),
-      );
-    }
-
-    if (showIntro) {
-      content = GameStartIntroOverlay(
-        onDismissed: () {
-          ref.read(gameIdsWithIntroShownProvider.notifier).markShown(game.id);
-        },
-        child: content,
-      );
-    }
-
-    if (!showIntro && game != null && pendingHerald != null) {
-      content = TribeFirstContactOverlay(
-        tribeName: pendingHerald.tribeName,
-        capitalName: pendingHerald.capitalName,
-        onDismissed: () {
-          ref
-              .read(tribeFirstContactHeraldsShownProvider.notifier)
-              .markShown(game.id, pendingHerald.tribeId);
-          ref.read(tribeFirstContactHeraldQueueProvider.notifier).dequeueHead();
-        },
-        child: content,
-      );
-    }
-
-    if (game != null && pendingDiplomacy != null) {
-      switch (pendingDiplomacy) {
-        case PendingDiplomacyOvertures(:final offers) when offers.isNotEmpty:
-          content = OvertureDialogueOverlay(
-            game: game,
-            pendingOvertures: offers,
-            onDecisions: (decisions) {
-              final service = ref.read(gameServiceProvider);
-              final orders = ref.read(currentOrdersProvider);
-              final result = service.resumeOvertureDecisions(
-                game,
-                offers,
-                decisions,
-                orders,
-              );
-              ref.read(turnResolutionResultApplierProvider).apply(result);
-            },
-            child: content,
-          );
-        case PendingDiplomacyIntervention(:final prompts)
-            when prompts.isNotEmpty:
-          content = InterventionDialogueOverlay(
-            game: game,
-            prompts: prompts,
-            onDecisions: (decisions) {
-              final service = ref.read(gameServiceProvider);
-              final orders = ref.read(currentOrdersProvider);
-              final result = service.resumeInterventionDecisions(
-                game,
-                decisions,
-                orders,
-              );
-              ref.read(turnResolutionResultApplierProvider).apply(result);
-            },
-            child: content,
-          );
-        case PendingDiplomacyCallToArms(:final pending) when pending.isNotEmpty:
-          content = CallToArmsDialogueOverlay(
-            game: game,
-            pending: pending,
-            onDecisions: (decisions) {
-              final service = ref.read(gameServiceProvider);
-              final orders = ref.read(currentOrdersProvider);
-              final result = service.resumeCallToArmsDecisions(
-                game,
-                decisions,
-                orders,
-              );
-              ref.read(turnResolutionResultApplierProvider).apply(result);
-            },
-            child: content,
-          );
-        case _:
-          break;
-      }
-    }
 
     return PopScope(
       canPop: false,
