+++ app/lib/features/game/widgets/production/production_panel_layouts.dart
@@ -1,0 +1,53 @@
+part of 'production_panel.dart';
+
+class _ProductionPanelNarrowLayout extends StatelessWidget {
+  const _ProductionPanelNarrowLayout({
+    super.key,
+    required this.availableSubpanel,
+    required this.allocationSubpanel,
+  });
+
+  final Widget availableSubpanel;
+  final Widget allocationSubpanel;
+
+  @override
+  Widget build(BuildContext context) {
+    return SingleChildScrollView(
+      padding: const EdgeInsets.all(CtSpacing.l),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.stretch,
+        children: [
+          availableSubpanel,
+          const SizedBox(height: 24),
+          allocationSubpanel,
+        ],
+      ),
+    );
+  }
+}
+
+class _ProductionPanelWideLayout extends StatelessWidget {
+  const _ProductionPanelWideLayout({
+    super.key,
+    required this.availableSubpanel,
+    required this.allocationSubpanel,
+  });
+
+  final Widget availableSubpanel;
+  final Widget allocationSubpanel;
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.all(CtSpacing.l),
+      child: Row(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Expanded(flex: 1, child: availableSubpanel),
+          const SizedBox(width: 24),
+          Expanded(flex: 2, child: allocationSubpanel),
+        ],
+      ),
+    );
+  }
+}
